+++ plugin-dev/Scripts/upload-debug-symbols.sh
@@ -54,16 +54,6 @@ if [ -z $CLI_LOG_LEVEL ]; then
     CLI_LOG_LEVEL="info"
 fi
 
-# Only check for specific platform if bEnableForAllTargetPlatforms is omitted or disabled
-ALL_TARGET_PLATFORMS_ENABLED_COUNT=$(grep -E -c "^bEnableForAllTargetPlatforms=[tT]rue$" "$DEFAULT_ENGINE_INI")
-if [ $ALL_TARGET_PLATFORMS_ENABLED_COUNT -eq 0 ]; then
-    TARGET_PLATFORM_ENTRY_COUNT=$(grep -E -c "^\+EnableTargetPlatforms=$targetPlatform$" "$DEFAULT_ENGINE_INI")
-    if [ $TARGET_PLATFORM_ENTRY_COUNT -eq 0 ]; then
-        echo "Sentry: Automatic symbols upload is disabled for build platform $targetPlatform. Skipping..."
-        exit
-    fi
-fi
-
 ENABLED_TARGETS=$(grep "EnableBuildTargets" "$DEFAULT_ENGINE_INI" | sed -n 's/^EnableBuildTargets=//p' | sed -e 's/^(\(.*\))$/\1/')
 if [ ! -z $ENABLED_TARGETS ]; then
     TARGETS_ARRAY=$(echo "$ENABLED_TARGETS" | sed -e 's/,/ /g')