+++ plugin-dev/Source/Sentry/Public/SentrySubsystem.h
@@ -319,9 +319,6 @@ private:
 	/** Check whether the event capturing should be enabled for the current build target */
 	bool IsCurrentBuildTargetEnabled();
 
-	/** Check whether the event capturing should be enabled for the current platform */
-	bool IsCurrentPlatformEnabled() const;
-
 	/** Check whether the event capturing should be enabled for promoted builds only */
 	bool IsPromotedBuildsOnlyEnabled();